build KEYBOARD_SEC_DOCK as module
sec_keyboard driver does not properly unregister the input device
when no dock is detected, which makes the system believe that a
physical keyboard is present at all times, which prevents
the software keyboard from being shown.

Some user own a Keyboard-Dock and missing the ability to use
the dock because it was disabled in kernel config.

Building it as module gives the user the ability to add a simple
init.d script to start the module at boot if they need it.

Change-Id: I2c2332d1293881b9ec88bf110c089b8bea8edfa0